insert confirmation dialog

+++ src/app/courses/containers/courses/courses.component.ts
@@ -6,6 +6,7 @@ import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/err
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -40,9 +41,17 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id).subscribe({
-      next: () => this.onSuccessSnackbar(),
-      error: () => this.onErrorSnackbar(),
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: course.name,
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.coursesService.remove(course._id).subscribe({
+          next: () => this.onSuccessSnackbar(),
+          error: () => this.onErrorSnackbar(),
+        });
+      }
     });
   }
